requirements.txt
aiogram==3.7.0
python-dotenv==1.0.1
google-genai==0.1.0
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
requests==2.32.3
pydantic==2.7.4
Flask>=2.3.0
gunicorn>=21.2.0
